refactor: gather the prior methods in the subpackage parameters
methods.py splits into parameters/methods.py (protocol and registry),
parametric.py, deep.py and _base.py (seed helpers and constraints). Priors,
intialize_priors and sample_from_priors move from simulations.py to
parameters/priors.py. LowerBound, UpperBound, DoubleBound and identity move from
utils.py to parameters/bijections.py; el.parameters re-exports the three
bounds.

Breaking: el.utils.LowerBound, UpperBound and DoubleBound are now
el.parameters.LowerBound, UpperBound and DoubleBound.
