--- src/elicito/methods.py
@@ -1,534 +1,0 @@
-"""
-Strategy objects for the prior-learning methods
-"""
-
-import contextlib
-from collections.abc import Iterator
-from typing import Any, Protocol
-
-import numpy as np
-import tensorflow as tf
-import tensorflow_probability as tfp  # type: ignore
-
-from elicito import networks
-from elicito.types import (
-    Initializer,
-    NFDict,
-    Parameter,
-    PriorMethods,
-)
-
-
-@contextlib.contextmanager
-def numpy_seed(seed: int) -> Iterator[None]:
-    """
-    Seed the numpy global generator for the duration of the block
-
-    A network can draw from the numpy global generator while it is built.
-    `elicito.networks.Permutation` does, and `tf.random.set_seed` does not
-    control that generator. Two networks built from the same seed then
-    permute differently, and a fitted network cannot be rebuilt from its
-    stored weights.
-
-    The generator of the caller is restored on exit.
-
-    Parameters
-    ----------
-    seed
-        Seed of the current workflow run.
-
-    Yields
-    ------
-    :
-        None. The block runs with the seeded generator.
-    """
-    state = np.random.get_state()  # noqa: NPY002
-    np.random.seed(seed)  # noqa: NPY002
-    try:
-        yield
-    finally:
-        np.random.set_state(state)  # noqa: NPY002
-
-
-def seed_pair(seed: int) -> Any:
-    """
-    Turn an integer seed into a stateless seed pair
-
-    A stateless pair makes every draw a function of the seed alone. The draws
-    then repeat without ``tf.random.set_seed``, which a compiled forward pass
-    cannot use: it clears the kernel caches of the whole graph.
-
-    ``tfp.random.split_seed`` accepts an integer, but it turns one into a pair
-    with a stateful ``tf.random.uniform``. That op branches on a value, which
-    a graph cannot do, and it would make the draws differ between calls.
-
-    Parameters
-    ----------
-    seed
-        Seed of the current workflow run.
-
-    Returns
-    -------
-    pair :
-        Seed as a pair of integers.
-
-    """
-    return tf.constant([0, seed], dtype=tf.int32)
-
-
-def _constraints(parameters: list[Parameter]) -> dict[str, Any]:
-    """Map each hyperparameter name to its constraint function."""
-    constraints: dict[str, Any] = {}
-    for param in parameters:
-        hyperparams = param["hyperparams"]
-        if hyperparams is None:
-            continue
-        for hyp in hyperparams:
-            constraints[hyperparams[hyp]["name"]] = hyperparams[hyp]["constraint"]
-    return constraints
-
-
-class PriorMethod(Protocol):
-    """Behaviour that differs between the prior-learning methods."""
-
-    name: str
-
-    def build(
-        self,
-        parameters: list[Parameter],
-        network: NFDict | None,
-        init_matrix_slice: dict[str, tf.Tensor] | None,
-        seed: int,
-    ) -> Any:
-        """Create the trainable prior object."""
-        ...
-
-    def sample(  # noqa: PLR0913
-        self,
-        initialized_priors: Any,
-        parameters: list[Parameter],
-        network: NFDict | None,
-        B: int,
-        num_samples: int,
-        seed: int,
-    ) -> Any:
-        """Draw prior samples of shape (B, num_samples, num_params)."""
-        ...
-
-    def trainable_variables(self, prior_model: Any) -> Any:
-        """Return the variables the optimizer updates."""
-        ...
-
-    def new_history(
-        self, prior_model: Any, parameters: list[Parameter]
-    ) -> dict[str, Any]:
-        """Create the per-epoch record, seeded with the initial values."""
-        ...
-
-    def record_epoch(
-        self,
-        history: dict[str, Any],
-        prior_sim: Any,
-        trainable_vars: Any,
-        parameters: list[Parameter],
-    ) -> None:
-        """Append this epoch's values to the record."""
-        ...
-
-    def finalize(
-        self,
-        res_ep: dict[str, Any],
-        output_res: dict[str, Any],
-        gradients_ep: Any,
-        trainable_vars: Any,
-    ) -> None:
-        """Add the method-specific entries to the results."""
-        ...
-
-    def check(
-        self,
-        parameters: list[Parameter],
-        network: NFDict | None,
-        initializer: Initializer | None,
-    ) -> None:
-        """Raise if the sections are not valid for this method."""
-        ...
-
-
-class ParametricPrior:
-    """Independent parametric priors."""
-
-    name = PriorMethods.parametric_prior.value
-
-    def __init__(self) -> None:
-        # cache filled by `new_history`; the map is constant during training
-        self._constraints: dict[str, Any] | None = None
-
-    def build(
-        self,
-        parameters: list[Parameter],
-        network: NFDict | None,
-        init_matrix_slice: dict[str, tf.Tensor] | None,
-        seed: int,
-    ) -> Any:
-        """Create the trainable prior object."""
-        # create dict with all hyperparameters
-        hyp_dict = dict()
-        hp_keys = list()
-        param_names = list()
-        hp_names = list()
-        initialized_hyperparam: dict[str, Any] = dict()
-
-        for i in range(len(parameters)):
-            hyperparameter = parameters[i]["hyperparams"]
-            if hyperparameter is not None:
-                num_hyperpar = len(hyperparameter)
-
-                hyp_dict[f"param{i}"] = hyperparameter
-                param_names += [parameters[i]["name"]] * num_hyperpar
-                hp_keys += list(hyperparameter.keys())
-                for j in range(num_hyperpar):
-                    current_key = list(hyperparameter.keys())[j]
-                    hp_names.append(hyperparameter[current_key]["name"])
-
-        checked_params = list()
-        for j, (i, hp_n, hp_k) in enumerate(
-            zip(tf.unique(param_names).idx, hp_names, hp_keys)
-        ):
-            if parameters[i]["hyperparams"] is not None:
-                hp_dict = parameters[i]["hyperparams"][hp_k]
-
-            if hp_dict is not None:
-                if hp_dict["shared"] and hp_dict["name"] in checked_params:
-                    pass
-                else:
-                    # get initial value
-                    if init_matrix_slice is not None:
-                        initial_value: Any = init_matrix_slice[hp_n]
-                    # initialize hyperparameter
-                    initialized_hyperparam[f"{hp_k}_{hp_n}"] = tf.Variable(
-                        initial_value=initial_value,
-                        trainable=True,
-                        name=f"{hp_dict['constraint_name']}.{hp_n}",
-                    )
-
-                    # save initialized priors
-                    init_prior = initialized_hyperparam
-
-                if hp_dict["shared"]:
-                    checked_params.append(hp_n)
-        return init_prior
-
-    def sample(  # noqa: D102, PLR0913
-        self,
-        initialized_priors: Any,
-        parameters: list[Parameter],
-        network: NFDict | None,
-        B: int,
-        num_samples: int,
-        seed: int,
-    ) -> Any:
-        # One stateless seed per parameter. A stateless seed repeats the draws
-        # without `tf.random.set_seed`, which a compiled forward pass cannot
-        # afford: it clears the kernel caches of the whole graph. One seed for
-        # every parameter would correlate the draws, so the seed is split.
-        seeds = tfp.random.split_seed(seed_pair(seed), n=len(parameters), salt="priors")
-        priors = []
-        for i in range(len(parameters)):
-            # get the prior distribution family as specified by the user
-            prior_family = parameters[i]["family"]
-
-            hp_k = list(parameters[i]["hyperparams"].keys())
-            init_dict = {}
-            for k in hp_k:
-                hp_n = parameters[i]["hyperparams"][k]["name"]
-                hp_constraint = parameters[i]["hyperparams"][k]["constraint"]
-                init_key = f"{k}_{hp_n}"
-                # init_dict[f"{k}"]=initialized_priors[init_key]
-                init_dict[f"{k}"] = hp_constraint(initialized_priors[init_key])
-            # sample from the prior distribution
-            priors.append(
-                prior_family(**init_dict).sample((B, num_samples), seed=seeds[i])
-            )
-        # stack all prior distributions into one tf.Tensor of
-        # shape (B, S, num_parameters)
-        if len(priors[0].shape) < 3:  # noqa: PLR2004
-            prior_samples = tf.stack(priors, axis=-1)
-        else:
-            prior_samples = tf.concat(priors, axis=-1)
-        return prior_samples
-
-    def trainable_variables(self, prior_model: Any) -> Any:
-        """Return the variables the optimizer updates."""
-        return prior_model.trainable_variables
-
-    def new_history(
-        self, prior_model: Any, parameters: list[Parameter]
-    ) -> dict[str, Any]:
-        """Create the per-epoch record, seeded with the initial values."""
-        constraints = _constraints(parameters)
-        self._constraints = constraints
-        history: dict[str, Any] = {}
-        for var in prior_model.trainable_variables:
-            name = var.name[:-2].split(".")[1]
-            history.setdefault(name, []).append(
-                float(constraints[name](var.numpy().copy()))
-            )
-        return history
-
-    def record_epoch(
-        self,
-        history: dict[str, Any],
-        prior_sim: Any,
-        trainable_vars: Any,
-        parameters: list[Parameter],
-    ) -> None:
-        """Create the per-epoch record, seeded with the initial values."""
-        constraints = self._constraints
-        if constraints is None:
-            constraints = _constraints(parameters)
-        for var in trainable_vars:
-            name = var.name[:-2].split(".")[1]
-            history[name].append(float(constraints[name](var.numpy().copy())))
-
-    def finalize(
-        self,
-        res_ep: dict[str, Any],
-        output_res: dict[str, Any],
-        gradients_ep: Any,
-        trainable_vars: Any,
-    ) -> None:
-        """Add the method-specific entries to the results."""
-        res_ep["hyperparameter_gradient"] = gradients_ep
-        output_res["learned_weights"] = {
-            f"weight_{i}": v.numpy().copy() for i, v in enumerate(trainable_vars)
-        }
-
-    def check(
-        self,
-        parameters: list[Parameter],
-        network: NFDict | None,
-        initializer: Initializer | None,
-    ) -> None:
-        """Raise if the sections are not valid for this method."""
-        if initializer is None:
-            msg = (
-                "If method is 'parametric_prior', "
-                " the section 'initializer' can't be None."
-            )
-            raise ValueError(msg)
-
-        if network is not None:
-            msg = (
-                "If method is 'parametric prior' "
-                "the 'network' is not used and should be set to None."
-            )
-            raise ValueError(msg)
-
-        # check that hyperparameter names are not redundant
-        hyp_names = []
-        hyp_shared = []
-        for i in range(len(parameters)):
-            if parameters[i]["hyperparams"] is None:
-                msg = (
-                    "When using method='parametric_prior', the argument "
-                    "'hyperparams' of el.parameter "
-                    "cannot be None."
-                )
-                raise ValueError(msg)
-
-            hyp_names.append(
-                [
-                    parameters[i]["hyperparams"][key]["name"]
-                    for key in parameters[i]["hyperparams"].keys()
-                ]
-            )
-            hyp_shared.append(
-                [
-                    parameters[i]["hyperparams"][key]["shared"]
-                    for key in parameters[i]["hyperparams"].keys()
-                ]
-            )
-        # flatten nested list
-        hyp_names_flat = sum(hyp_names, [])  # noqa: RUF017
-        hyp_shared_flat = sum(hyp_shared, [])  # noqa: RUF017
-
-        hyperparams = initializer["hyperparams"]
-        # exact values win over a method string, as resolve_init_method does
-        if hyperparams is not None:
-            for k in hyperparams:
-                if k not in hyp_names_flat:
-                    msg = (
-                        f"Hyperparameter name '{k}' doesn't "
-                        "match any name specified in the parameters "
-                        "section. Have you misspelled the name?"
-                    )
-                    raise ValueError(msg)
-
-        seen = []
-        duplicate = []
-        share = []
-        for n, s in zip(hyp_names_flat, hyp_shared_flat):
-            if n not in seen:
-                seen.append(n)
-            elif s:
-                share.append(n)
-            else:
-                duplicate.append(n)
-
-        if len(duplicate) != 0:
-            msg = (
-                "The following hyperparameter have the same "
-                f"name but are not shared: {duplicate}. \n"
-                "Have you forgot to set shared=True?"
-            )
-            raise ValueError(msg)
-
-
-class DeepPrior:
-    """Joint non-parametric prior via a normalizing flow."""
-
-    name = PriorMethods.deep_prior.value
-
-    def build(
-        self,
-        parameters: list[Parameter],
-        network: NFDict | None,
-        init_matrix_slice: dict[str, tf.Tensor] | None,
-        seed: int,
-    ) -> Any:
-        """Create the trainable prior object."""
-        # for more information see BayesFlow documentation
-        # https://bayesflow.org/api/bayesflow.inference_networks.html
-        if network is not None:
-            INN = network["inference_network"]
-
-            # The permutation layers draw from the numpy global generator,
-            # see `numpy_seed`.
-            with numpy_seed(seed):
-                invertible_neural_network = INN(**network["network_specs"])  # type: ignore [call-arg]
-
-            # save initialized priors
-            init_prior = invertible_neural_network
-
-            # build network
-            # initialize base distribution
-            base_dist = network["base_distribution"](num_params=len(parameters))  # type: ignore
-            # the base distribution never changes during training;
-            # keep it on the prior object instead of rebuilding it per epoch
-            init_prior.base_distribution = base_dist
-            # sample from base distribution
-            u = base_dist.sample((128, 200))
-            init_prior(u, None)
-        return init_prior
-
-    def sample(  # noqa: D102, PLR0913
-        self,
-        initialized_priors: Any,
-        parameters: list[Parameter],
-        network: NFDict | None,
-        B: int,
-        num_samples: int,
-        seed: int,
-    ) -> Any:
-        # reuse the base distribution built in `build`. The seed is stateless,
-        # see `ParametricPrior.sample`.
-        u = initialized_priors.base_distribution.sample(
-            (B, num_samples),
-            seed=tfp.random.split_seed(seed_pair(seed), n=1, salt="base")[0],
-        )
-        # apply transformation function to samples from base distr.
-        (unconstr_priors, _) = initialized_priors(u, condition=None, inverse=False)
-        # apply parameter constraints if specified
-        constr_priors = []
-        for j in range(len(parameters)):
-            constr = parameters[j]["constraint"]
-            constr_priors.append(constr(unconstr_priors[:, :, j]))
-        prior_samples = tf.stack(constr_priors, axis=-1)
-        return prior_samples
-
-    def trainable_variables(self, prior_model: Any) -> Any:
-        """Return the variables the optimizer updates."""
-        return prior_model.init_priors.trainable_variables
-
-    def new_history(
-        self, prior_model: Any, parameters: list[Parameter]
-    ) -> dict[str, Any]:
-        """Create the per-epoch record, seeded with the initial values."""
-        return {"means": [], "stds": []}
-
-    def record_epoch(
-        self,
-        history: dict[str, Any],
-        prior_sim: Any,
-        trainable_vars: Any,
-        parameters: list[Parameter],
-    ) -> None:
-        """Create the per-epoch record, seeded with the initial values."""
-        history["means"].append(tf.reduce_mean(prior_sim, (0, 1)))
-        history["stds"].append(tf.reduce_mean(tf.math.reduce_std(prior_sim, 1), 0))
-
-    def finalize(
-        self,
-        res_ep: dict[str, Any],
-        output_res: dict[str, Any],
-        gradients_ep: Any,
-        trainable_vars: Any,
-    ) -> None:
-        """Add the method-specific entries to the results."""
-        output_res["num_NN_weights"] = [v.shape for v in trainable_vars]
-        output_res["learned_weights"] = {
-            f"weight_{i}": v.numpy().copy() for i, v in enumerate(trainable_vars)
-        }
-
-    def check(
-        self,
-        parameters: list[Parameter],
-        network: NFDict | None,
-        initializer: Initializer | None,
-    ) -> None:
-        """Raise if the sections are not valid for this method."""
-        if network is None:
-            msg = "If method is 'deep prior',  the section 'network' can't be None."
-            raise ValueError(msg)
-
-        if initializer is not None:
-            msg = (
-                "For method 'deep_prior' the "
-                "'initializer' is not used and should be set to None."
-            )
-            raise ValueError(msg)
-
-        if network["network_specs"]["num_params"] != len(parameters):
-            msg = (
-                "The number of model parameters as "
-                "specified in the parameters section, must match the "
-                "number of parameters specified in the network."
-                f"Expected {len(parameters)} but got "
-                f"{network['network_specs']['num_params']}"
-            )
-            raise ValueError(msg)
-
-        if network["base_distribution"].__class__ != networks.BaseNormal:
-            msg = (
-                "Currently only the standard normal distribution "
-                "is implemented as base distribution. "
-                "See GitHub issue #35."
-            )
-            raise NotImplementedError(msg)
-
-
-_METHODS: dict[str, type[PriorMethod]] = {
-    ParametricPrior.name: ParametricPrior,
-    DeepPrior.name: DeepPrior,
-}
-
-
-def get_method(name: str) -> PriorMethod:
-    """Return a new strategy object for a ``trainer["method"]`` string."""
-    try:
-        method_cls = _METHODS[name]
-    except KeyError:
-        msg = f"Unknown method {name!r}. Valid: {sorted(_METHODS)}."
-        raise ValueError(msg) from None
-    return method_cls()
