+++ src/elicito/parameters/__init__.py
@@ -1,0 +1,17 @@
+"""
+Prior distributions of the model parameters, and the methods that learn them
+"""
+
+from elicito.parameters import bijections, deep, methods, parametric, priors
+from elicito.parameters.bijections import DoubleBound, LowerBound, UpperBound
+
+__all__ = [
+    "DoubleBound",
+    "LowerBound",
+    "UpperBound",
+    "bijections",
+    "deep",
+    "methods",
+    "parametric",
+    "priors",
+]
